refactor: renamed FrontendProjectsTable to NodeProjectsTable (used on both frontend and backend)
